Complete the hero section
Use Display:flex on the whole section, so that the image is on one
side and the text, subtext and button are on the other side. Then add
width to the text section, so that it wraps where needs to be and
occupies a determined space on the hero. The style is also taken from
the tutorial document.